Fix subprocess streaming logs in jupyter notebook by watching PIPE
